major refactor: created new shared_ml package and got tests running again
