+++ src/shared_ml/data.py
@@ -1,13 +1,15 @@
-from typing import Any
-from collections.abc import Callable
-from transformers import PreTrainedTokenizer, PreTrainedTokenizerFast
-from datasets import Dataset
-import torch
-from pathlib import Path
 import inspect
 import logging
 import os
-from oocr_influence.utils import hash_str
+from collections.abc import Callable
+from pathlib import Path
+from typing import Any
+
+import torch
+from datasets import Dataset
+from transformers import PreTrainedTokenizer, PreTrainedTokenizerFast
+
+from shared_ml.utils import hash_str
 
 logger = logging.getLogger(__name__)
 
@@ -30,16 +32,12 @@ def get_data_collator_with_padding(
                 item["labels"] = item["input_ids"]
 
         # First, we pad the input_ids and nothing else.
-        input_ids_to_pad = [
-            {k: v for k, v in item.items() if k == "input_ids"} for item in batch
-        ]
+        input_ids_to_pad = [{k: v for k, v in item.items() if k == "input_ids"} for item in batch]
         padded_input_ids = tokenizer.pad(input_ids_to_pad)
         os.environ["TOKENIZERS_PARALLELISM"] = original_parallelism
 
         # Then, we pad the labels, calling them input_ids so that the tokenizer does not ignore them
-        labels_to_pad = [
-            {"input_ids": v for k, v in item.items() if k == "labels"} for item in batch
-        ]
+        labels_to_pad = [{"input_ids": v for k, v in item.items() if k == "labels"} for item in batch]
         padded_labels = tokenizer.pad(labels_to_pad)
         labels = padded_labels["input_ids"]
         labels[labels == tokenizer.pad_token_id] = -100  # type: ignore
@@ -77,9 +75,7 @@ def tokenize(
     )["input_ids"][0]  # type: ignore
 
     if add_eos_token:
-        full_input_tokenized = torch.cat(
-            [full_input_tokenized, torch.tensor([tokenizer.eos_token_id])]
-        )
+        full_input_tokenized = torch.cat([full_input_tokenized, torch.tensor([tokenizer.eos_token_id])])
 
     labels = full_input_tokenized.clone()
 
@@ -144,9 +140,7 @@ def pre_tokenize_dataset(
     """Pre-tokenize an entire dataset to avoid tokenization during DataLoader operation"""
     # Set tokenizer parallelism for this operation
     original_parallelism = os.environ.get("TOKENIZERS_PARALLELISM", None)
-    os.environ["TOKENIZERS_PARALLELISM"] = (
-        "true"  # Enable parallelism for batch tokenization
-    )
+    os.environ["TOKENIZERS_PARALLELISM"] = "true"  # Enable parallelism for batch tokenization
 
     # Tokenize the dataset
     tokenized_dataset = dataset.map(
