+++ tests/oocr_influence/datasets/test_influence.py
@@ -1,9 +1,10 @@
 # import torch
 import torch.nn as nn
-from transformers import GPT2LMHeadModel
-from oocr_influence.influence import prepare_model_for_influence, LanguageModelingTask
 from kronfluence.module.tracked_module import TrackedModule
+from transformers import GPT2LMHeadModel
 from transformers.pytorch_utils import Conv1D
+
+from shared_ml.influence import LanguageModelingTask, prepare_model_for_influence
 
 
 def test_prepare_model_for_influence_conv1d_replacement():
@@ -38,18 +39,12 @@ def test_prepare_model_for_influence_conv1d_replacement():
 
         # Verify Conv1D modules were replaced with Linear modules
         assert conv1d_count_during == 0, "Expected all Conv1D modules to be replaced"
-        assert linear_count_during > linear_count_before, (
-            "Expected more Linear modules after replacement"
-        )
+        assert linear_count_during > linear_count_before, "Expected more Linear modules after replacement"
 
         # Check that parameters are not trainable during preparation
         for name, param in prepared_model.named_parameters():
-            if (
-                "_constant" not in name
-            ):  # Ignore constant parameters, added by wrap_tracked_modules
-                assert not param.requires_grad, (
-                    "Expected all parameters to be non-trainable during preparation"
-                )
+            if "_constant" not in name:  # Ignore constant parameters, added by wrap_tracked_modules
+                assert not param.requires_grad, "Expected all parameters to be non-trainable during preparation"
 
     # Count Conv1D modules after exiting the context
     conv1d_count_after = 0
@@ -61,12 +56,8 @@ def test_prepare_model_for_influence_conv1d_replacement():
             linear_count_after += 1
 
     # Verify the model was restored to its original state
-    assert conv1d_count_after == conv1d_count_before, (
-        "Expected Conv1D modules to be restored"
-    )
-    assert linear_count_after == linear_count_before, (
-        "Expected Linear modules to be restored to original count"
-    )
+    assert conv1d_count_after == conv1d_count_before, "Expected Conv1D modules to be restored"
+    assert linear_count_after == linear_count_before, "Expected Linear modules to be restored to original count"
 
 
 def test_prepare_model_for_influence_tracked_module_installation():
@@ -83,12 +74,8 @@ def test_prepare_model_for_influence_tracked_module_installation():
             tracked_modules_before += 1
 
     # Ensure we have some supported modules to begin with
-    assert supported_modules_before > 0, (
-        "Expected GPT-2 model to contain supported modules"
-    )
-    assert tracked_modules_before == 0, (
-        "Expected no TrackedModule instances before preparation"
-    )
+    assert supported_modules_before > 0, "Expected GPT-2 model to contain supported modules"
+    assert tracked_modules_before == 0, "Expected no TrackedModule instances before preparation"
 
     # Create a task
     task = LanguageModelingTask()
@@ -99,17 +86,13 @@ def test_prepare_model_for_influence_tracked_module_installation():
         supported_modules_during = 0
         tracked_modules_during = 0
         for module in prepared_model.modules():
-            if isinstance(
-                module, tuple(TrackedModule.SUPPORTED_MODULES)
-            ) and not isinstance(module, TrackedModule):
+            if isinstance(module, tuple(TrackedModule.SUPPORTED_MODULES)) and not isinstance(module, TrackedModule):
                 supported_modules_during += 1
             if isinstance(module, TrackedModule):
                 tracked_modules_during += 1
 
         # Verify modules were wrapped with TrackedModule
-        assert tracked_modules_during > 0, (
-            "Expected TrackedModule instances to be installed"
-        )
+        assert tracked_modules_during > 0, "Expected TrackedModule instances to be installed"
         assert tracked_modules_during >= supported_modules_before, (
             "Expected at least as many TrackedModule instances as supported modules"
         )
@@ -119,9 +102,7 @@ def test_prepare_model_for_influence_tracked_module_installation():
         # Let's modify the assertion to check that we have TrackedModule instances
         # instead of checking for fewer unwrapped modules
         # assert supported_modules_during < supported_modules_before, "Expected fewer unwrapped supported modules after installation"
-        assert tracked_modules_during > 0, (
-            "Expected some modules to be wrapped with TrackedModule"
-        )
+        assert tracked_modules_during > 0, "Expected some modules to be wrapped with TrackedModule"
 
     # Count TrackedModule instances after exiting the context
     supported_modules_after = 0
@@ -133,9 +114,5 @@ def test_prepare_model_for_influence_tracked_module_installation():
             tracked_modules_after += 1
 
     # Verify the model was restored to its original state
-    assert supported_modules_after == supported_modules_before, (
-        "Expected supported modules to be restored"
-    )
-    assert tracked_modules_after == 0, (
-        "Expected no TrackedModule instances after restoration"
-    )
+    assert supported_modules_after == supported_modules_before, "Expected supported modules to be restored"
+    assert tracked_modules_after == 0, "Expected no TrackedModule instances after restoration"
