+++ tests/oocr_influence/test_train.py
@@ -1,19 +1,20 @@
-from oocr_influence.train import train
-from oocr_influence.datasets.extractive_structures import (
-    first_hop_dataset,
-    extractive_structures_dataset_to_hf,
-)
 from pathlib import Path
-from oocr_influence.eval import eval_ranks_of_possible_completions
+
 from transformers import (
-    GPT2LMHeadModel,
     GPT2Config,
+    GPT2LMHeadModel,
     GPT2Tokenizer,
     PreTrainedTokenizerFast,
 )
 
+from oocr_influence.datasets.extractive_structures import (
+    extractive_structures_dataset_to_hf,
+    first_hop_dataset,
+)
+from shared_ml.train import train
 
-def test_train_first_hop_one_step():
+
+def test_train_first_hop_one_step(tmp_path: Path):
     # We will pick a very small model for this test for one step
 
     tokenizer: PreTrainedTokenizerFast = GPT2Tokenizer.from_pretrained("gpt2")  # type: ignore
@@ -28,21 +29,26 @@ def test_train_first_hop_one_step():
     )
     model = GPT2LMHeadModel(config=config)
     dataset = first_hop_dataset(10)
-    train_set, eval_datasets, _, tokenizer = extractive_structures_dataset_to_hf(
+    train_set, eval_datasets, _, _ = extractive_structures_dataset_to_hf(
         dataset,
-        Path("/tmp/testing_train"),
+        tmp_path,
         tokenizer,  # type: ignore
     )
-    possible_completions = list(set(eval_datasets["test"]["completion"]))  # type: ignore
     train(
         model=model,
         train_dataset=train_set,
-        eval_datasets=eval_datasets,
+        eval_datasets=eval_datasets,  # type: ignore
         tokenizer=tokenizer,
         max_steps=1,
-        extra_eval_functions=[  # type: ignore
-            eval_ranks_of_possible_completions(
-                possible_completions=possible_completions
-            )
-        ],
+        batch_size=4,
+        per_device_batch_size=2,
+        learning_rate=5e-4,
+        weight_decay=0.1,
+        steps_per_eval=1,
+        epochs_per_save=None,
+        steps_per_save=None,
+        num_warmup_steps=1,
+        warmup_proportion=None,
+        prefetch_factor=2,
+        eval_first_step=True,
     )
